mission backend update

+++ T-Manager/src/main/java/cpt202/groupwork/repository/MissionRepository.java
@@ -9,7 +9,7 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MissionRepository extends JpaRepository<Mission, Integer> {
+public interface MissionRepository extends JpaRepository<Mission, Long> {
 //  List<TodoRepository> findByUsername(String username);
   /**
    * 通过 projectId 找到 List<Mission>
